Calori calculation has been started.
